--- Microsoft.Azure.Cosmos/src/Headers/ResponseHeaders.cs
@@ -1,180 +1,0 @@
-//------------------------------------------------------------
-// Copyright (c) Microsoft Corporation.  All rights reserved.
-//------------------------------------------------------------
-
-namespace Microsoft.Azure.Cosmos
-{
-    using System;
-    using System.Collections;
-    using System.Collections.Generic;
-    using System.Globalization;
-    using System.Linq;
-    using System.Reflection;
-    using Microsoft.Azure.Documents;
-
-    /// <summary>
-    /// HTTP headers in a <see cref="ResponseMessage"/>.
-    /// </summary>
-    public class ResponseHeaders : HeadersBase, IEnumerable
-    {
-        [CosmosKnownHeaderAttribute(HeaderName = HttpConstants.HttpHeaders.RequestCharge)]
-        private string requestCharge
-        {
-            get
-            {
-                return this.RequestCharge == 0 ? null : this.RequestCharge.ToString();
-            }
-            set
-            {
-                this.RequestCharge = string.IsNullOrEmpty(value) ? 0 : double.Parse(value, CultureInfo.InvariantCulture);
-            }
-        }
-
-        /// <summary>
-        /// Gets the request charge for this request from the Azure Cosmos DB service.
-        /// </summary>
-        /// <value>
-        /// The request charge measured in request units.
-        /// </value>
-        public virtual double RequestCharge { get; internal set; }
-
-        /// <summary>
-        /// Gets the activity ID for the request from the Azure Cosmos DB service.
-        /// </summary>
-        /// <value>
-        /// The activity ID for the request.
-        /// </value>
-        [CosmosKnownHeaderAttribute(HeaderName = HttpConstants.HttpHeaders.ActivityId)]
-        public virtual string ActivityId { get; internal set; }
-
-        /// <summary>
-        /// Gets the entity tag associated with the resource from the Azure Cosmos DB service.
-        /// </summary>
-        /// <value>
-        /// The entity tag associated with the resource.
-        /// </value>
-        /// <remarks>
-        /// ETags are used for concurrency checking when updating resources. 
-        /// </remarks>
-        [CosmosKnownHeaderAttribute(HeaderName = HttpConstants.HttpHeaders.ETag)]
-        public virtual string ETag { get; internal set; }
-
-        /// <summary>
-        /// Gets the Continuation Token in the current <see cref="ResponseMessage"/>.
-        /// </summary>
-        [CosmosKnownHeaderAttribute(HeaderName = HttpConstants.HttpHeaders.Continuation)]
-        public virtual string Continuation { get; internal set; }
-
-        /// <summary>
-        /// Gets the Content Type for the current content in the <see cref="ResponseMessage"/>.
-        /// </summary>
-        [CosmosKnownHeaderAttribute(HeaderName = HttpConstants.HttpHeaders.ContentType)]
-        public virtual string ContentType { get; internal set; }
-
-        /// <summary>
-        /// Gets the Session Token for the current <see cref="ResponseMessage"/>.
-        /// </summary>
-        /// <remarks>
-        /// Session Token is used along with Session Consistency.
-        /// </remarks>
-        [CosmosKnownHeaderAttribute(HeaderName = HttpConstants.HttpHeaders.SessionToken)]
-        public virtual string Session { get; internal set; }
-
-        /// <summary>
-        /// Gets the Content Length for the current content in the <see cref="ResponseMessage"/>.
-        /// </summary>
-        [CosmosKnownHeaderAttribute(HeaderName = HttpConstants.HttpHeaders.ContentLength)]
-        public virtual string ContentLength
-        {
-            get
-            {
-                return this.contentLength;
-            }
-            set
-            {
-                this.ContentLengthAsLong = value != null ? long.Parse(value, CultureInfo.InvariantCulture) : 0;
-                this.contentLength = value;
-            }
-        }
-
-        /// <summary>
-        /// Gets the Location for the current content in the <see cref="ResponseMessage"/>.
-        /// </summary>
-        [CosmosKnownHeaderAttribute(HeaderName = HttpConstants.HttpHeaders.Location)]
-        public virtual string Location { get; internal set; }
-
-        internal TimeSpan? RetryAfter;
-
-        internal SubStatusCodes SubStatusCode = SubStatusCodes.Unknown;
-
-        internal long ContentLengthAsLong;
-
-        [CosmosKnownHeaderAttribute(HeaderName = WFConstants.BackendHeaders.SubStatus)]
-        internal string SubStatusCodeLiteral
-        {
-            get
-            {
-                return this.subStatusCodeLiteral;
-            }
-            set
-            {
-                this.SubStatusCode = ResponseHeaders.GetSubStatusCodes(value);
-                this.subStatusCodeLiteral = value;
-            }
-        }
-
-        [CosmosKnownHeaderAttribute(HeaderName = HttpConstants.HttpHeaders.RetryAfterInMilliseconds)]
-        internal string RetryAfterLiteral
-        {
-            get
-            {
-                return this.retryAfterInternal;
-            }
-            set
-            {
-                this.RetryAfter = ResponseHeaders.GetRetryAfter(value);
-                this.retryAfterInternal = value;
-            }
-        }
-
-        private string contentLength;
-
-        private string subStatusCodeLiteral;
-
-        private string retryAfterInternal;
-
-        private static KeyValuePair<string, PropertyInfo>[] knownHeaderProperties = CosmosMessageHeadersInternal.GetHeaderAttributes<ResponseHeaders>();
-
-        internal override Dictionary<string, CosmosCustomHeader> CreateKnownDictionary()
-        {
-            return ResponseHeaders.knownHeaderProperties.ToDictionary(
-                    knownProperty => knownProperty.Key,
-                    knownProperty => new CosmosCustomHeader(
-                            () => (string)knownProperty.Value.GetValue(this),
-                            (string value) => { knownProperty.Value.SetValue(this, value); }), 
-                    StringComparer.OrdinalIgnoreCase);
-        }
-
-        internal static SubStatusCodes GetSubStatusCodes(string value)
-        {
-            uint nSubStatus = 0;
-            if (uint.TryParse(value, NumberStyles.Integer, CultureInfo.InvariantCulture, out nSubStatus))
-            {
-                return (SubStatusCodes)nSubStatus;
-            }
-
-            return SubStatusCodes.Unknown;
-        }
-
-        internal static TimeSpan? GetRetryAfter(string value)
-        {
-            long retryIntervalInMilliseconds = 0;
-            if (long.TryParse(value, NumberStyles.Number, CultureInfo.InvariantCulture, out retryIntervalInMilliseconds))
-            {
-                return TimeSpan.FromMilliseconds(retryIntervalInMilliseconds);
-            }
-
-            return null;
-        }
-    }
-}
