+++ Microsoft.Azure.Cosmos/src/Resource/QueryResponses/ReadFeedResponse.cs
@@ -11,7 +11,7 @@ namespace Microsoft.Azure.Cosmos
     {
         protected ReadFeedResponse(
             IEnumerable<T> resource,
-            ResponseHeaders responseMessageHeaders,
+            Headers responseMessageHeaders,
             bool hasMoreResults)
             : base(
                 httpStatusCode: HttpStatusCode.Accepted,
@@ -35,7 +35,7 @@ namespace Microsoft.Azure.Cosmos
         }
 
         internal static ReadFeedResponse<TInput> CreateResponse<TInput>(
-            ResponseHeaders responseMessageHeaders,
+            Headers responseMessageHeaders,
             Stream stream,
             CosmosSerializer jsonSerializer,
             bool hasMoreResults)
@@ -54,7 +54,7 @@ namespace Microsoft.Azure.Cosmos
         }
 
         internal static ReadFeedResponse<TInput> CreateResponse<TInput>(
-            ResponseHeaders responseMessageHeaders,
+            Headers responseMessageHeaders,
             IEnumerable<TInput> resources,
             bool hasMoreResults)
         {
