+++ Microsoft.Azure.Cosmos/src/Resource/Scripts/StoredProcedureResponse.cs
@@ -27,7 +27,7 @@ namespace Microsoft.Azure.Cosmos.Scripts
         /// </summary>
         internal StoredProcedureResponse(
            HttpStatusCode httpStatusCode,
-           ResponseHeaders headers,
+           Headers headers,
            StoredProcedureProperties storedProcedureProperties)
             : base(
                httpStatusCode,
